+++ pareg_core/src/err/arg_error.rs
@@ -113,6 +113,11 @@ impl ArgError {
     pub fn kind(&self) -> &ArgErrKind {
         &self.0.kind
     }
+    
+    pub fn map_ctx(mut self, f: impl FnOnce(&mut ArgErrCtx)) -> Self {
+        f(&mut self.0);
+        self
+    }
 }
 
 impl std::error::Error for ArgError {